refactor: rename app folder
