+++ apps/client/src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
-import App from './app2/App.tsx';
+import App from './app/App.tsx';
 import './i18n.tsx';
 
 createRoot(document.querySelector('#root') || document.body).render(
